move to anonymous namespace
